requirements.txt
opencv-python>=4.8.0
easyocr>=1.7.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
pyserial>=3.5
flask>=3.0.0
